dsc-it-100: Brute force installer code
This is my first rust program!

It opens /dev/ttyUSB0 and proceeds to brute force
the installer pin code.  It should take somewhat less than
8 hours to run depending on how many retries/etc it needs
and how numerically low the code is.

Funny enough, my panel (a 1575) seems really
sensitive to having commands/etc sent to quickly
and gets itself into "stupid" mode where it acknowledges
the commands but never acts on them until its
hard reset. The early versions of this program were using
status commands/etc as part of a fancy state machine but it
turned out that it wasn't as robust as just waiting fixed
amounts of time between sending each keypress, and a fixed
amount of time for the panel to stop responding.

Signed-off-by: Jeremy Linton <lintonrjeremy@gmail.com>